+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+* Fix `scaffold` command for empty class/modules
 * Type check method definition with empty body
 * Add `STDOUT` and `StringIO` minimal definition
 * Fix validate command to load stdlib